优化杀死app流程

+++ baselibrary/src/main/java/com/aibabel/baselibrary/base/BaseApplication.java
@@ -236,16 +236,19 @@ public abstract class BaseApplication extends Application {
             @Override
             public void onActivityCreated(Activity activity, Bundle savedInstanceState) {
                 activityLinkedList.add(activity);
+                canExit=false;
             }
 
             @Override
             public void onActivityStarted(Activity activity) {
                 Logger.d(activity.getLocalClassName());
                 stateCount++;
+                canExit=false;
             }
 
             @Override
             public void onActivityResumed(Activity activity) {
+                canExit=false;
             }
 
             @Override
@@ -268,19 +271,37 @@ public abstract class BaseApplication extends Application {
             }
         });
     }
-
+    private  static  volatile boolean canExit=true;
     /**
      * 退出所有app
      */
     public static void exit() {
         for (Activity activity : activityLinkedList) {
             activity.finish();
+
         }
-        new Handler().postDelayed(new Runnable() {
+        canExit=true;
+        new Thread(new Runnable() {
             @Override
             public void run() {
-                android.os.Process.killProcess(android.os.Process.myPid());
+                try {
+                    Thread.sleep(3000);
+                    Log.e("canExit===",String.valueOf(canExit));
+                    if ( canExit){
+                        android.os.Process.killProcess(android.os.Process.myPid());
+                    }
+                } catch (InterruptedException e) {
+                    e.printStackTrace();
+                }
+
             }
-        },1500);
+        }).start();
+//        handler=new Handler();
+//        handler.postDelayed(new Runnable() {
+//            @Override
+//            public void run() {
+//                android.os.Process.killProcess(android.os.Process.myPid());
+//            }
+//        },1500);
     }
 }
